Remove fallback on default

+++ core/platform/cache/redis_cache_services.py
@@ -42,6 +42,8 @@ class RedisClient:
             return
 
         redishost = redis_services.get_redis_host()
+        if not redishost:
+            return
         self._oppia_redis_client = redis.StrictRedis(
             host=redishost,
             port=feconf.REDISPORT,
@@ -55,24 +57,22 @@ class RedisClient:
         )
         self._is_client_initialized = True
 
-    def get_oppia_redis_client(self) -> redis.StrictRedis[str]:
+    def get_oppia_redis_client(self) -> Optional[redis.StrictRedis[str]]:
         """Initializes redis model and obtains oppia redis client.
 
         Returns:
             redis.StrictRedis[str]. The oppia redis client.
         """
         self._initialize_client()
-        assert self._oppia_redis_client is not None
         return self._oppia_redis_client
 
-    def get_cloud_ndb_redis_client(self) -> redis.StrictRedis[str]:
+    def get_cloud_ndb_redis_client(self) -> Optional[redis.StrictRedis[str]]:
         """Initializes redis model and obtains cloud ndb redis client.
 
         Returns:
             redis.StrictRedis[str]. The cloud ndb redis client.
         """
         self._initialize_client()
-        assert self._cloud_ndb_redis_client is not None
         return self._cloud_ndb_redis_client
 
 
@@ -90,6 +90,8 @@ def get_memory_cache_stats() -> caching_domain.MemoryCacheStats:
         keys stored as values.
     """
     oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
+    if not oppia_redis_client:
+        return caching_domain.MemoryCacheStats(0, 0, 0)
     redis_full_profile = oppia_redis_client.memory_stats()
     return caching_domain.MemoryCacheStats(
         redis_full_profile['total.allocated'],
@@ -100,10 +102,12 @@ def get_memory_cache_stats() -> caching_domain.MemoryCacheStats:
 
 def flush_caches() -> None:
     """Wipes the Redis caches clean."""
-    oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
-    oppia_redis_client.flushdb()
-    cloud_ndb_redis_client = REDIS_CLIENT.get_cloud_ndb_redis_client()
-    cloud_ndb_redis_client.flushdb()
+    if oppia_redis_client:
+        oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
+        oppia_redis_client.flushdb()
+    if cloud_ndb_redis_client:
+        cloud_ndb_redis_client = REDIS_CLIENT.get_cloud_ndb_redis_client()
+        cloud_ndb_redis_client.flushdb()
 
 
 def get_multi(keys: List[str]) -> List[Optional[str]]:
@@ -118,6 +122,8 @@ def get_multi(keys: List[str]) -> List[Optional[str]]:
     """
     assert isinstance(keys, list)
     oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
+    if not oppia_redis_client:
+        return []
     return oppia_redis_client.mget(keys)
 
 
@@ -134,6 +140,8 @@ def set_multi(key_value_mapping: Dict[str, str]) -> bool:
     """
     assert isinstance(key_value_mapping, dict)
     oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
+    if not oppia_redis_client:
+        return False
     return oppia_redis_client.mset(key_value_mapping)
 
 
@@ -148,5 +156,7 @@ def delete_multi(keys: List[str]) -> int:
     """
     for key in keys:
         assert isinstance(key, str)
+    if not oppia_redis_client:
+        return 0
     oppia_redis_client = REDIS_CLIENT.get_oppia_redis_client()
     return oppia_redis_client.delete(*keys)
